requirements.txt
flask
python-dotenv
replicate 
flask-cors
